+++ pcs/common/reports/types.py
@@ -1,4 +1,4 @@
-from typing import NewType
+from typing import Collection, NewType
 
 AddRemoveContainerType = NewType("AddRemoveContainerType", str)
 AddRemoveItemType = NewType("AddRemoveItemType", str)
@@ -6,6 +6,7 @@ BoothConfigUsedWhere = NewType("BoothConfigUsedWhere", str)
 DefaultAddressSource = NewType("DefaultAddressSource", str)
 FenceHistoryCommandType = NewType("FenceHistoryCommandType", str)
 ForceCode = NewType("ForceCode", str)
+ForceFlags = Collection[ForceCode]
 MessageCode = NewType("MessageCode", str)
 DeprecatedMessageCode = NewType("DeprecatedMessageCode", str)
 PcsCommand = NewType("PcsCommand", str)