+++ pcs/lib/commands/resource.py
@@ -92,6 +92,7 @@ from pcs.lib.resource_agent import (
     resource_agent_error_to_report_item,
     split_resource_agent_name,
 )
+from pcs.lib.sbd_stonith import ensure_some_stonith_remains
 from pcs.lib.tools import get_tmp_cib
 from pcs.lib.validate import ValueTimeInterval
 from pcs.lib.xml_tools import (
@@ -352,6 +353,24 @@ _find_bundle = partial(
 )
 
 
+def _are_meta_disabled(meta_attributes: Mapping[str, str]) -> bool:
+    return meta_attributes.get("target-role", "Started").lower() == "stopped"
+
+
+def _can_be_evaluated_as_positive_num(value: str) -> bool:
+    string_wo_leading_zeros = str(value).lstrip("0")
+    return bool(string_wo_leading_zeros) and (
+        string_wo_leading_zeros[0] in list("123456789")
+    )
+
+
+def _is_clone_deactivated_by_meta(meta_attributes: Mapping[str, str]) -> bool:
+    return _are_meta_disabled(meta_attributes) or any(
+        not _can_be_evaluated_as_positive_num(meta_attributes.get(key, "1"))
+        for key in ["clone-max", "clone-node-max"]
+    )
+
+
 def create(  # noqa: PLR0913
     env: LibraryEnvironment,
     resource_id: str,
@@ -427,8 +446,7 @@ def create(  # noqa: PLR0913
         wait,
         [resource_id],
         _ensure_disabled_after_wait(
-            ensure_disabled
-            or resource.common.are_meta_disabled(meta_attributes)
+            ensure_disabled or _are_meta_disabled(meta_attributes)
         ),
         required_cib_version=get_required_cib_version_for_primitive(
             operation_list
@@ -573,8 +591,8 @@ def create_as_clone(  # noqa: PLR0913
         [resource_id],
         _ensure_disabled_after_wait(
             ensure_disabled
-            or resource.common.are_meta_disabled(meta_attributes)
-            or resource.common.is_clone_deactivated_by_meta(clone_meta_options)
+            or _are_meta_disabled(meta_attributes)
+            or _is_clone_deactivated_by_meta(clone_meta_options)
         ),
         required_cib_version=get_required_cib_version_for_primitive(
             operation_list
@@ -697,8 +715,7 @@ def create_in_group(  # noqa: PLR0913
         wait,
         [resource_id],
         _ensure_disabled_after_wait(
-            ensure_disabled
-            or resource.common.are_meta_disabled(meta_attributes)
+            ensure_disabled or _are_meta_disabled(meta_attributes)
         ),
         required_cib_version=get_required_cib_version_for_primitive(
             operation_list
@@ -853,8 +870,7 @@ def create_into_bundle(  # noqa: PLR0913
         wait,
         [resource_id],
         _ensure_disabled_after_wait(
-            ensure_disabled
-            or resource.common.are_meta_disabled(meta_attributes)
+            ensure_disabled or _are_meta_disabled(meta_attributes)
         ),
         required_cib_version=required_cib_version,
     ) as resources_section:
@@ -955,8 +971,7 @@ def bundle_create(  # noqa: PLR0913
         wait,
         [bundle_id],
         _ensure_disabled_after_wait(
-            ensure_disabled
-            or resource.common.are_meta_disabled(meta_attributes)
+            ensure_disabled or _are_meta_disabled(meta_attributes)
         ),
         required_cib_version=(
             Version(3, 2, 0) if container_type == "podman" else None
@@ -1041,8 +1056,7 @@ def bundle_reset(  # noqa: PLR0913
         wait,
         [bundle_id],
         _ensure_disabled_after_wait(
-            ensure_disabled
-            or resource.common.are_meta_disabled(meta_attributes)
+            ensure_disabled or _are_meta_disabled(meta_attributes)
         ),
         # The only requirement for CIB schema version currently is:
         #   if container_type == "podman" then required_version = '3.2.0'
@@ -1170,27 +1184,6 @@ def bundle_update(  # noqa: PLR0913
         )
 
 
-def _disable_validate_and_edit_cib(
-    env: LibraryEnvironment,
-    cib: _Element,
-    resource_or_tag_ids: StringCollection,
-) -> List[_Element]:
-    resource_el_list, report_list = _find_resources_expand_tags(
-        cib, resource_or_tag_ids
-    )
-    env.report_processor.report_list(report_list)
-    if env.report_processor.report_list(
-        _resource_list_enable_disable(
-            resource_el_list,
-            resource.common.disable,
-            IdProvider(cib),
-            env.get_cluster_state(),
-        )
-    ).has_errors:
-        raise LibraryError()
-    return resource_el_list
-
-
 def _disable_get_element_ids(
     disabled_resource_el_list: Iterable[_Element],
 ) -> Tuple[Set[str], Set[str]]:
@@ -1259,6 +1252,7 @@ def disable(
     env: LibraryEnvironment,
     resource_or_tag_ids: StringCollection,
     wait: WaitType = False,
+    force_flags: reports.types.ForceFlags = (),
 ):
     """
     Disallow specified resources to be started by the cluster
@@ -1277,7 +1271,42 @@ def disable(
         )
 
     wait_timeout = env.ensure_wait_satisfiable(wait)
-    _disable_validate_and_edit_cib(env, env.get_cib(), resource_or_tag_ids)
+    cib = env.get_cib()
+    resource_el_list, report_list = _find_resources_expand_tags(
+        cib, resource_or_tag_ids
+    )
+    env.report_processor.report_list(report_list)
+
+    if any(
+        resource.stonith.is_stonith(resource_el)
+        for resource_el in resource_el_list
+    ):
+        env.report_processor.report_list(
+            ensure_some_stonith_remains(
+                env,
+                get_resources(cib),
+                [str(res.attrib["id"]) for res in resource_el_list],
+                sbd_being_disabled=False,
+                force_flags=force_flags,
+            )
+        )
+
+    # Validation done, do the disabling. Do not mind errors that happened so
+    # far. The disabling may report errors on its own and we want the user to
+    # see those. In case of errors, we exit before pushing CIB, not making any
+    # change to cluster configuration.
+    env.report_processor.report_list(
+        _resource_list_enable_disable(
+            resource_el_list,
+            resource.common.disable,
+            IdProvider(cib),
+            env.get_cluster_state(),
+        )
+    )
+
+    if env.report_processor.has_errors:
+        raise LibraryError()
+
     _push_cib_wait(
         env,
         wait_timeout,
@@ -1319,9 +1348,11 @@ def disable_safe(
 
     wait_timeout = env.ensure_wait_satisfiable(wait)
     cib = env.get_cib()
-    resource_el_list = _disable_validate_and_edit_cib(
-        env, cib, resource_or_tag_ids
+    resource_el_list, report_list = _find_resources_expand_tags(
+        cib, resource_or_tag_ids
     )
+    env.report_processor.report_list(report_list)
+
     if any(
         resource.stonith.is_stonith(resource_el)
         for resource_el in resource_el_list
@@ -1333,6 +1364,22 @@ def disable_safe(
                 )
             )
         )
+
+    # Validation done, do the disabling. Do not mind errors that happened so
+    # far. The disabling may report errors on its own and we want the user to
+    # see those. In case of errors, we exit before pushing CIB, not making any
+    # change to cluster configuration.
+    env.report_processor.report_list(
+        _resource_list_enable_disable(
+            resource_el_list,
+            resource.common.disable,
+            IdProvider(cib),
+            env.get_cluster_state(),
+        )
+    )
+    if env.report_processor.has_errors:
+        raise LibraryError()
+
     disabled_resource_id_set, inner_resource_id_set = _disable_get_element_ids(
         resource_el_list
     )
@@ -1390,9 +1437,27 @@ def disable_simulate(
         )
 
     cib = env.get_cib()
-    resource_el_list = _disable_validate_and_edit_cib(
-        env, cib, resource_or_tag_ids
+    resource_el_list, report_list = _find_resources_expand_tags(
+        cib, resource_or_tag_ids
     )
+    env.report_processor.report_list(report_list)
+
+    # Validation done, do the disabling. Do not mind errors that happened so
+    # far. The disabling may report errors on its own and we want the user to
+    # see those. In case of errors, we exit before pushing CIB, not making any
+    # change to cluster configuration.
+    env.report_processor.report_list(
+        _resource_list_enable_disable(
+            resource_el_list,
+            resource.common.disable,
+            IdProvider(cib),
+            env.get_cluster_state(),
+        )
+    )
+
+    if env.report_processor.has_errors:
+        raise LibraryError()
+
     disabled_resource_id_set, inner_resource_id_set = _disable_get_element_ids(
         resource_el_list
     )
@@ -1459,11 +1524,14 @@ def enable(
 
 
 def _resource_list_enable_disable(
-    resource_el_list, func, id_provider, cluster_state
-):
+    resource_el_list: Iterable[_Element],
+    func: Callable[[_Element, IdProvider], None],
+    id_provider: IdProvider,
+    cluster_state,
+) -> ReportItemList:
     report_list = []
     for resource_el in resource_el_list:
-        res_id = resource_el.attrib["id"]
+        res_id = str(resource_el.attrib["id"])
         try:
             if not is_resource_managed(cluster_state, res_id):
                 report_list.append(