+++ pcs_test/tier0/lib/commands/resource/test_resource_create.py
@@ -22,6 +22,46 @@ from pcs_test.tools.misc import outdent
 TIMEOUT = 10
 
 
+class AreMetaDisabled(TestCase):
+    def test_detect_is_disabled(self):
+        self.assertTrue(resource._are_meta_disabled({"target-role": "Stopped"}))
+        self.assertTrue(resource._are_meta_disabled({"target-role": "stopped"}))
+
+    def test_detect_is_not_disabled(self):
+        self.assertFalse(resource._are_meta_disabled({}))
+        self.assertFalse(resource._are_meta_disabled({"target-role": "any"}))
+
+
+class IsCloneDeactivatedByMeta(TestCase):
+    def assert_is_disabled(self, meta_attributes):
+        self.assertTrue(resource._is_clone_deactivated_by_meta(meta_attributes))
+
+    def assert_is_not_disabled(self, meta_attributes):
+        self.assertFalse(
+            resource._is_clone_deactivated_by_meta(meta_attributes)
+        )
+
+    def test_detect_is_disabled(self):
+        self.assert_is_disabled({"target-role": "Stopped"})
+        self.assert_is_disabled({"target-role": "stopped"})
+        self.assert_is_disabled({"clone-max": "0"})
+        self.assert_is_disabled({"clone-max": "00"})
+        self.assert_is_disabled({"clone-max": 0})
+        self.assert_is_disabled({"clone-node-max": "0"})
+        self.assert_is_disabled({"clone-node-max": "abc1"})
+
+    def test_detect_is_not_disabled(self):
+        self.assert_is_not_disabled({})
+        self.assert_is_not_disabled({"target-role": "any"})
+        self.assert_is_not_disabled({"clone-max": "1"})
+        self.assert_is_not_disabled({"clone-max": "01"})
+        self.assert_is_not_disabled({"clone-max": 1})
+        self.assert_is_not_disabled({"clone-node-max": "1"})
+        self.assert_is_not_disabled({"clone-node-max": 1})
+        self.assert_is_not_disabled({"clone-node-max": "1abc"})
+        self.assert_is_not_disabled({"clone-node-max": "1.1"})
+
+
 def create(  # noqa: PLR0913
     env,
     *,