+++ pcs_test/tier1/cib_resource/test_stonith_enable_disable.py
@@ -1,6 +1,10 @@
 from pcs_test.tier1.cib_resource.common import ResourceTest
 from pcs_test.tools.bin_mock import get_mock_settings
 
+ERRORS_HAVE_OCCURRED = (
+    "Error: Errors have occurred, therefore pcs is unable to continue\n"
+)
+
 
 class Enable(ResourceTest):
     def setUp(self):
@@ -74,8 +78,31 @@ class Disable(ResourceTest):
                 </primitive>
             </resources>""",
         )
-        self.assert_effect(
+        self.assert_pcs_fail(
             "stonith disable S".split(),
+            (
+                "Error: Requested action lefts the cluster with no enabled "
+                "means to fence nodes, resulting in the cluster not being able "
+                "to recover from certain failure conditions, use --force to "
+                "override\n" + ERRORS_HAVE_OCCURRED
+            ),
+        )
+
+    def test_disable_enabled_stonith_forced(self):
+        self.assert_effect(
+            "stonith create S fence_pcsmock_minimal".split(),
+            """<resources>
+                <primitive class="stonith" id="S" type="fence_pcsmock_minimal">
+                    <operations>
+                        <op id="S-monitor-interval-60s" interval="60s"
+                            name="monitor"
+                        />
+                    </operations>
+                </primitive>
+            </resources>""",
+        )
+        self.assert_effect(
+            "stonith disable S --force".split(),
             """<resources>
                 <primitive class="stonith" id="S" type="fence_pcsmock_minimal">
                     <meta_attributes id="S-meta_attributes">
@@ -90,6 +117,11 @@ class Disable(ResourceTest):
                     </operations>
                 </primitive>
             </resources>""",
+            stderr_full=(
+                "Warning: Requested action lefts the cluster with no enabled "
+                "means to fence nodes, resulting in the cluster not being able "
+                "to recover from certain failure conditions\n"
+            ),
         )
 
     def test_keep_disabled_stonith(self):