+++ pcs_test/tier1/legacy/test_stonith.py
@@ -1380,7 +1380,23 @@ class StonithTest(TestCase, AssertPcsMixin):
         )
 
         self.pcs_runner.corosync_conf_opt = None
-        self.assert_pcs_success("stonith delete test_stonith".split())
+        self.assert_pcs_fail(
+            "stonith delete test_stonith".split(),
+            (
+                "Error: Requested action lefts the cluster with no enabled "
+                "means to fence nodes, resulting in the cluster not being able "
+                "to recover from certain failure conditions, use --force to "
+                "override\n" + ERRORS_HAVE_OCCURRED
+            ),
+        )
+        self.assert_pcs_success(
+            "stonith delete test_stonith --force".split(),
+            stderr_full=(
+                "Warning: Requested action lefts the cluster with no enabled "
+                "means to fence nodes, resulting in the cluster not being able "
+                "to recover from certain failure conditions\n"
+            ),
+        )
 
         self.pcs_runner.corosync_conf_opt = self.temp_corosync_conf.name
         self.assert_pcs_success(